moving some functions from pint.utils to pint.derived_quantities.  Moved associated tests to tests/test_derived_quantities.  Made related changes elsewhere
